extract common heartbeatslist component and apply to all heartbeat list screens
